move OptimoveConfig up, open access to affected package-private methods
